FIX.#30
J'ai ajouté des tests pour reproduire les chemins du code dans Lucca, et c'est bon, ils sont couverts.
